MenuItem (replicator v0.1.0)

A replicator menu item.

Types

charge_amount()

Specs

charge_amount() :: non_neg_integer()

The charge amount for an item.

The amount of replicator charge it costs to replicate a given item. Must be greater than 0.

t()

Specs

t() :: %MenuItem{energy_cost: charge_amount(), icon: String.t()}

A menu item.

Properties

enenergy_cost: The amount of replicator charge it costs to replicate one of the item. icon: An icon to represent the item, ideally an emoji.

Functions

new(energy_cost, icon)

Create a new %MenuItem{...}.

Parameters

enenergy_cost: The amount of replicator charge it costs to replicate one of the item.

icon: An icon to represent the item, ideally an emoji.

Returns

Returns a new %MenuItem{...}. Returns nil if passed invalid parameters.