PBSReplicator (replicator v0.1.0)

Why cook when you can use a replicator! Sadly this module can't provide a real replicator, but it can at least give you a simulated digital one. 🙂

A Virtual Star Trek-style Replicator

Replicators have a charge that gets exhausted by replicating food. Replicators can be recharged.

All replicators share a single menu which is accessible and updatable via functions.

Types

t()

Specs

t() :: %PBSReplicator{charge: MenuItem.charge_amount(), menu: Map.t()}

Functions

add_menu_item(replicator, arg)

Add a MenuItem to the replicator's menu.

Parammeters:

replicator - The replicator to recharge.

{name, icon, energy_cost} - A Tuple of the name, icon, and energy cost of the new item.

Sample:

iex> {:ok, rep} = PBSReplicator.add_menu_item(rep, {:tako, "🐙", 9})
{:ok,
 %PBSReplicator{
   charge: 225,
   menu: %{
     pancakes: %MenuItem{energy_cost: 11, icon: "🥞"},
     popcorn: %MenuItem{energy_cost: 1, icon: "🍿"},
     tako: %MenuItem{energy_cost: 9, icon: "🐙"},
     waffles: %MenuItem{energy_cost: 15, icon: "🧇"}
   }
 }}

charge(_, _)

get_charge(replicator)

Get the replicator's current charge level.

Parammeters:

replicator - The replicator to check.

Sample:

iex> {:ok, charge} = PBSReplicator.get_charge(rep)
{:ok, 150}

get_menu(replicator)

Get the menu of foods supported by the replicator.

Parammeters:

replicator - The replicator to get the menu from.

Sample:

iex> {:ok, menu} = PBSReplicator.get_menu(rep)
{:ok,
 %{
   pancakes: %MenuItem{energy_cost: 11, icon: "🥞"},
   popcorn: %MenuItem{energy_cost: 1, icon: "🍿"},
   waffles: %MenuItem{energy_cost: 15, icon: "🧇"}
 }}

new()

Specs

new() :: {:ok, Map.t()}

Create a default-charged Replicator.

Sample:

iex> {:ok, rep} = PBSReplicator.new()
{:ok,
 %PBSReplicator{
   charge: 100,
   menu: %{
     pancakes: %MenuItem{energy_cost: 11, icon: "🥞"},
     popcorn: %MenuItem{energy_cost: 1, icon: "🍿"},
     waffles: %MenuItem{energy_cost: 15, icon: "🧇"}
   }
 }}

new(initial_charge)

Specs

new(MenuItem.charge_amount()) :: {atom(), Map.t()}

Create a new Replicator charged to a given amount.

Replicators default to an initial charge of 100, but an alternative initial charge may be passed.

Parammeters:

initial_charge (default: 100) - The replicator's initial charge.

Sample:

iex> {:ok, rep} = PBSReplicator.new(150)
{:ok,
 %PBSReplicator{
   charge: 150,
   menu: %{
     pancakes: %MenuItem{energy_cost: 11, icon: "🥞"},
     popcorn: %MenuItem{energy_cost: 1, icon: "🍿"},
     waffles: %MenuItem{energy_cost: 15, icon: "🧇"}
   }
 }}

recharge(replicator, charge_amount)

Recharge the replicator's current charge level.

Parammeters:

replicator - The replicator to recharge.

charge_amount - The amount to charge.

Sample:

iex> {:ok, rep} = PBSReplicator.recharge(rep, 75)
{:ok,
 %PBSReplicator{
   charge: 225,
   menu: %{
     pancakes: %MenuItem{energy_cost: 11, icon: "🥞"},
     popcorn: %MenuItem{energy_cost: 1, icon: "🍿"},
     waffles: %MenuItem{energy_cost: 15, icon: "🧇"}
   }
 }}

replicate(replicator, arg)

Make some food from the menu.

Parammeters:

replicator - The replicator to use.

{name, count} - A Tuple of the name of the item to make and the number of items (default: 1).

Returns

Returns the Tuple {:ok, replicator, food} if successful, {:error, reason} if fails.

Sample:

iex> {:ok, rep, food} = PBSReplicator.replicate(rep, {:tako, 3})
{:ok, %PBSReplicator{charge: 198, ...]}, "🐙🐙🐙"}